+++ functions/api/events/index.ts
@@ -1,0 +1,6 @@
+
+export async function onRequest({ request, env }) {
+    return new Response("Hello from the Events API", {
+        headers: { "Content-Type": "text/plain" },
+    });
+}
