+++ functions/api/index.ts
@@ -1,0 +1,6 @@
+
+export async function onRequest({ request, env }) {
+    return new Response("Hello from the API", {
+        headers: { "Content-Type": "text/plain" },
+    });
+}
